+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 - Standalone presenter action template is not analysed if presenter dir and template dir are siblings
+- Method call __toString() is considered template render call
 
 ## [0.16.0] - 2023-08-24
 ### Changed